advanced yield section extracted
