+++ apps/earn-protocol/features/strategy-details/components/StrategyDetailsAdvancedYield/mapper.tsx
